Backend/requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
requests==2.32.5
python-dotenv==1.2.1
joblib==1.5.2
